Fix clippy warnings for Rust 1.93 stable
- Replace manual Default impls with #[derive(Default)] where equivalent
- Remove unreachable wildcard arms in exhaustive matches
- Collapse nested if statements
- Remove no-effect * 1 operations
- Fix unwrap-after-is_some with if-let
- Use iterator enumerate instead of index loop
- Convert while-match to while-let
- Fix empty line after doc comment (use //! module docs)
- Allow type_complexity and too_many_arguments in Bevy crates
- Use struct init with ..Default::default() instead of field reassignment

+++ crates/cc_client/src/renderer/fog.rs
@@ -81,8 +81,8 @@ pub fn init_fog(
     map_res: Res<MapResource>,
     mut materials: ResMut<Assets<ColorMaterial>>,
 ) {
-    let w = map_res.map.width as u32;
-    let h = map_res.map.height as u32;
+    let w = map_res.map.width;
+    let h = map_res.map.height;
     let size = (w * h) as usize;
     fog.width = w;
     fog.height = h;
